Introduce `BoTorchModel` to OSS (#374)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/374

This PR introduces a simple way to create a custom BoTorch model in Ax.

The new modular `BoTorchModel` object is composed of a choice of a `Surrogate` class and a choice of an `Acquisition` class. The `Surrogate` is a wrapper around a BoTorch GP `Model` and the `Acquisition` is a wrapper around a BoTorch `AcquisitionFunction`. These `BoTorchModel`-s are also storable with JSON and SQL.

Note that the functionality in this commit is under construction and not yet finalized.

A work-in-progress tutorial for the new setup is also part of this commit, as 'tutorials/botorch_modular.ipynb'.

--------

Reviewed By: lena-kashtelyan

Differential Revision: D23376065

fbshipit-source-id: 3059c64208193c3d807030ff8acb57cab3a9e73e