Refactor ref_point into optimization_config as syntactic sugar around outcome_constraints
Summary:
Add ref_point as a syntactic sugar argument for OptimizationConfig's constructor, and as a property of OptimizationConfig.  Translate it to/from OutcomeConstraints in OptimizationConfig.

Not yet refactoring ReferencePoint into its own object - I think it doesn't have much functionality and it wouldn't simplify the API at all (but I'm open to changing my mind).

Reviewed By: Balandat

Differential Revision: D23697215

fbshipit-source-id: 3c4ded446ce50dbe535c951264dd98bb04cf28b2