Allow surrogate options in `BoTorchModel`
Summary:
Part 1 if what was previously D23883304 (split for ease of review).

In some cases, we want to pass surrogate options and not a whole surrogate (for instance, if we want to pass just some one kwarg to the underlying `Model` but don't want to go through instantiation of custom `Surrogate` for it). This allows that approach and is utilized in D23883304.

Reviewed By: Balandat

Differential Revision: D24001292

fbshipit-source-id: 6413c672ed6271cf793c65dab6e222a97338c1f0