Don't re-fetch metrics unnecessarily
Summary:
This diff aims to address the issue described here: https://fb.quip.com/8EZ4AKhZ9d7q#FQKACAqDqd2.

After a fair amount of discussion with ldworkin, we decided that it was cleanest to implement this alternative behavior on the metric classes (as they depend on a class property of the metric and to avoid overly complicated logic in `experiment.fetch_data`).

Reviewed By: Balandat

Differential Revision: D23626518

fbshipit-source-id: 2bfbe90f93c7f3ba9191c82923ea7010367db016