Make `compute_model_dependencies` an instance method instead of class method
Summary:
We don't leverage the fact that `compute_model_dependencies` is a class method anywhere, and it's unnecessarily limiting (for instance, if you need to do something fancy in your `CustomAcquisition.__init__` and then need to let `compute_model_dependencies` have access to the results of that computation, you can't currently as `compute_model_dependencies` is a class method).

While keeping `compute_model_dependencies` a class method is somewhat more functional & less object-oriented, I think in this case we can forego that, as D23593010 shows that making  `compute_model_dependencies`  a class method necessitates passing around extra kwargs –– at that point, its worthwhile to just have it be an instance method, I believe.

Reviewed By: sdsingh

Differential Revision: D24225692

fbshipit-source-id: 6c94cc5a80affed6b73c84ec6b4b4a5552d30d6a