Fix _gen_multiple()
Summary: In the current implementation, we will ignore the provided `num_generator_runs` and will generate something on the order of `max_parallelism` in some cases (even for `gen(n=1)`). This causes us to create more generator runs that are saved to the GenerationStrategy then we return to the user, causing sobol inconsistencies.

Reviewed By: lena-kashtelyan

Differential Revision: D24401120

fbshipit-source-id: e5b8e7d7d94692eca00bed8f6906c175dd20445b