add test to check for init files
Summary:
added a test in ax/ax/utils/testing as per lena-kashtelyan suggestion. It walks through the ax/ax directory and makes sure that there is at least on __init__.py file in each directory.

pretty sure I was told that I wouldn't need the library, but couldn't get it working without. Let me know if there is a different way that I should be doing this!

Reviewed By: sdsingh

Differential Revision: D24552183

fbshipit-source-id: 65e816057d677ee73d64c95dcea4c8189d8d9394