`_skip_experiment_save` option to scheduler
Summary:
In `Scheduler.from_stored_experiment`, we do the following:
- load experiment and GS,
- `__init__` a new `Scheduler` with the above experiment and GS.

In the course of the `__init__`, `Scheduler` automatically saves experiment and GS (this is necessary because 1) scheduler updates the experiment slightly on `__init__`, 2) it's never guaranteed that the user did not forget to save experiment after making the last modification to it prior to running a scheduler with it).

This new flag allows to skip experiment resaving, useful if:
1) we just reloaded experiment and are `__init__`-ing the scheduler as part of `Scheduler.from_stored_experiment`,
2) for a user who really knows what they are doing and wants to skip saving because they just saved (hopefully the leading underscore in `_skip_experiment_save` will deter those who are uncertain of what they are doing).

Specifically, this is **very** needed now because when reloading large experiments, the subsequent unnecessary re-save breaks the Bento kernel in non-deterministic ways (ldworkin and I believe that the cause is running out of memory due to all the expensive operations the save performs). Normally we save experiment incrementally –– after each update to it, we just save relevant trials –– but in this case where we just save whole exp at once, Bento doesn't seem to handle it well.

Reviewed By: Balandat

Differential Revision: D24580210

fbshipit-source-id: 2408c2c8a9b3d1cd74f4ce7da8f9453e1da9b631