Set parameters in constraints to search space' parameters
Summary:
In our parameter constraints, we currently accept `Parameter`-s as arguments, here for example: https://fburl.com/diffusion/uphwd24k. Those parameters are expected to be copies of the parameters already in search space, and we validate that: https://fburl.com/diffusion/ul0cnmwu.

This diff reassigns the parameters in constraint object to the the existing search space parameters to:
1) avoid keeping around copies of parameters,
2) align with our reloading behavior, where we point constraints to the parameters in search space (we don't store the copies) *,
3) avoid issues with setting db_id on the parameters, since the parameter copies in constraints can't have db id-s (since they are not stored).

 ---
(*) in `Decoder`, we pass parameters to `parameter_constraint_from_sqa` and put those into the newly recreated constraints: https://fburl.com/diffusion/rso1wxvb, in `Encoder` we just don't store the parameters that are part of the constraint: https://fburl.com/diffusion/psfkdalx

Reviewed By: sdsingh

Differential Revision: D24594866

fbshipit-source-id: e1e94fb64da74a21f735e8af449f69a10bce7aa7