7/n: Set `db_id` on user-facing classes in saving functions
Summary:
NOTE: All diffs in this stack will need to land together. Logic is interdependent, and I'm splitting the diffs for ease of review. It's not possible to split these perfectly such that all tests pass on each diff, so bear with me please : )

In this diff, we actually set `db_id` on user-facing classes during `save_experiment`

Reviewed By: ldworkin

Differential Revision: D24601318

fbshipit-source-id: 8ccfdbe3f874ad72ecc3a957adcfa60c910755bc