5/n: Adjust encoding functions to return `obj_to_sqa` where needed
Summary:
NOTE: All diffs in this stack will need to land together. Logic is interdependent, and I'm splitting the diffs for ease of review. It's not possible to split these perfectly such that all tests pass on each diff, so bear with me please : )

In this diff, I adjust encoding functions to return `obj_to_sqa` lists, which we can then use to set `db_id` on user-facing classes.

Reviewed By: ldworkin

Differential Revision: D24601191

fbshipit-source-id: a8c7750b243fa858b0defe27ef72b46224e5940a