6/n: Adjust decoding functions to set db_ids on user-facing classes upon re-creation
Summary:
NOTE: All diffs in this stack will need to land together. Logic is interdependent, and I'm splitting the diffs for ease of review. It's not possible to split these perfectly such that all tests pass on each diff, so bear with me please : )

Specifically what is done in this diff I plan on replacing with a decorator once all these initial changes are in. I don't want to overcomplicate the work now by introducing that decorator –– I believe it will be easier to figure one out once `db_id` is correctly used and all tests are passing.

Reviewed By: ldworkin

Differential Revision: D24601249

fbshipit-source-id: 57f7ad421b8fe878e1755e3f14829edbaefa1ed2