Validate data in service API, including for SQ in `AxBatchClient`
Summary:
Preventing issues like the one in T79485119 coming up later on in cryptics ways, when in fact we should just require SQ data to be present in data for any batch that has the SQ set.
keskici, this validation should work in both Maximus and TRIP, correct? In the latter I know it must, since TRIP uses relative GP (which requires SQ data).

Reviewed By: ldworkin

Differential Revision: D24922390

fbshipit-source-id: c415b5229d30c3e0c88d5f2ae95847419da0e14c