Pareto frontier computation skips the check for a trial_index when data is provided
Summary:
Existing validation of batch trials requires a trial_index to be provided; when data is not explicitly provided, trial_index is used to fetch data to fit the MOO model.

This change removes that check when data is explicitly given as trial_index is not required for this purpose. (Trial index is also used to compute the status_quo_prediction.)

Reviewed By: ldworkin

Differential Revision: D25063112

fbshipit-source-id: ae66dc7f1ad01c075be70f4c1a230c1c82aac647