Error in generation strategy if using update logic with metrics that are available while running
Summary: Using `model.update` instead of refitting the model in generation strategy currently relies on using trial completion times to identify new data (model is always updated with data for trials that became `COMPLETED` since last update). I have not yet had a chance to think though how this should work with metrics that are available while trials are running, since those can have new data before being `COMPLETED`. I'd prefer to just explicitly raise an error in that case for now, so we can sort out that use case when we have an example application, with which in mind we can think the case through.

Reviewed By: ldworkin

Differential Revision: D25040260

fbshipit-source-id: 709561cd954366991da95ee3125c8dd01ba2051d