Support dataclass in annotation checks
Summary:
Dataclass decorators cause issues with our platform 009 due to the "fully_annotated" test failing. Basically, `dataclass` sets a bunch of helper methods on classes (e.g. `__repl__`) that aren't fully annotated in the Py3 source. Because the decorated classes themselves are in Ax, we treat these `__repl__` methods are part of Ax and fail the test.

Banning `dataclass` could work. But the other fix is to make the test ignore these weird functions. The one fix I found was too check the source of every func; if it isn't in Ax (i.e. some weird lambda setter like in `dataclass`),  should raise an error on `getsource()`

Reviewed By: ldworkin

Differential Revision: D25911287

fbshipit-source-id: fe2d5880b089bc4abcf0e90349946fe26eda8076