Fix SobolGenerator tests (#485)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/485

These tests are failing due to https://github.com/pytorch/pytorch/pull/49710, which changed the behavior of the pytorch SobolEngine. The distributional properties are arguably better due to the new direction numbers, but there is a breaking change to the realized samples.

This changes the tests to not rely on the specific values generated by the upstream sampler, making the tests less brittle and work with both the latest pytorch release and the current master.

Reviewed By: ldworkin

Differential Revision: D26207619

fbshipit-source-id: adc7480f9a47bee3b64404b811975953481068ea