Create versioned_sidebars and versioned_docs dir (#488)
Summary:
Goal is to prevent errors of [this form](https://github.com/facebook/Ax/runs/1828238043) that pop up transiently (but frequently) during our nightly cron latest website push, which generate oncall noise.

We shouldn't need to create these directories ourselves, but there is a race condition in the Docusaurus code. This line [here](https://fburl.com/gp8xx5i4) which creates the directory returns a promise, and we don't wait for it to resolve.

Pull Request resolved: https://github.com/facebook/Ax/pull/488

Test Plan:
we'll see if this works https://github.com/facebook/Ax/actions/runs/540860533 but of course even if it does, given that this is a race condition, there's no way of truly knowing ...

update, it worked

Reviewed By: lena-kashtelyan

Differential Revision: D26277423

Pulled By: ldworkin

fbshipit-source-id: 3312c3ae48cfbc0145843374c49d5b97524a74bc