Add a Yeo-Johnson power transform to Ax (#502)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/502

This adds a Yeo-Johnson power transform to Ax. I have modeled this based of `InverseGaussianCdfY` as these are very similar except for the difference that the power transform can be inverted. We could also have used Box-Cox, but sklearn only implements the one-parameter version which can only be used for positive values, while Yeo-Johnson handles any inputs.

I currently use CI matching to untransform the values, but there are some issues with this. The first one is that not all values can be untransformed, so we need to clip the values back in the range that we can invert (similar to `InverseGaussianCdfY`). The second issue is that outliers tend to get massive error bars when they are untransformed because of the fact that the derivative of the inverse tends to be very large  there. Thing behave very well in the transformed space though which is what I'm most interested in shadowing.

A couple of things that would be great to get some feedback on:
- I use the  `PowerTransformer` in Sklearn to fit the transform. This means that I'm holding on to (rather simple) Sklearn objects that I need to transform and untransform. I'm not sure if this causes any issues with serialization.
- I currently don't touch off-diagonal covariance entries, similarly to `InverseGaussianCdfY`. I assume these are mostly present in field experiments and not AutoML, but please let me know what to do about that. Some better input checking probably makes sense there.

Reviewed By: Balandat

Differential Revision: D26383524

fbshipit-source-id: 31d5e4109e9239e0b5c844677d31f8d8e6e43d45