Exclude OOD candidates from clamping
Summary:
Currently, when untransforming candidates, we also clamp them all to the search space. This is usually find *unless* the candidates returned are OOD. Which can be the case for TS, among other models.  This can cause unintended transformations that break generation (e.g. N492337).

To handle this, we can disable clamping on models that fit on OOD training data. Should give us good coverage on this issue

Reviewed By: bletham

Differential Revision: D26780629

fbshipit-source-id: cfd2a2273ceddc56ab9cdd57df77a59b278d4704