add scalarized outcome constraint
Summary:
Motivation: Support the case that the constraints are weighted sum of some metric outcomes. For instance, we want no increase in cpu cost across a certain region. The outcomes are observed at country level. The constraint is then represented as `\sum w_c * cpu_c` for c in the region e.g. altoona.

Implementation:
- add a new class `ScalarizedOutcomeConstraint` (similar to the `ScalarizedObjective`) that takes  a list of metrics and the corresponding weight as input.
- to pass validation in optimization config, add a instance check and only validate `OutcomeConstraint`
- change `extract_outcome_constraints` to time the weight to the metric component of each constraint.

Reviewed By: ldworkin

Differential Revision: D26517747

fbshipit-source-id: e564a3a083e26e8964bd513e83b54cbd8c43277f