Fix bug in modelbridge pareto utils
Summary:
Fixes a bug in the modelbridge utils used for computing the Pareto front. In the current code, observations are converted into tensors using whatever metric order happened to be in the observation, and not the metric order that is used by the modelbridge. These are in general different because observations are not required to have even the same set of metrics as the modelbridge (not all metrics are always observed for every data point).

This fixes the bug by ensuring that when we convert observation data to a tensor it uses the order specified by the modelbridge.

Impact: The order used by the modelbridge will always be sorted. So if the metrics on the experiment were not in alphabetical order, it means the tensors were being constructed basically with shuffled columns. That is, the Y and Yvar passed into `pareto_frontier_evaluator` here had the columns shuffled:
https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/ax/ax/models/torch/botorch_moo_defaults.py?commit=d810ee9ee2c95b94de46e5245d4b62cdb7f20a23&lines=218

Importantly, objective_weights, objective_thresholds, and outcome_constraints did not have their columns shuffled; they had the correct ordering specified by the modelbridge. This means that outcome constraints and objective thresholds were being applied to the wrong metrics. Pareto fronts have thus been incorrect for any experiment that (1) uses objective thresholds or outcome constraints, and (2) didn't happen to have the metrics in alphabetical order.

Note that this affects only pareto fronts and hypervolumes computed using modelbridge_utils.pareto_frontier . The pareto fronts computed in the plot code uses a totally separate code path that isn't affected.

Reviewed By: 2timesjay

Differential Revision: D26930981

fbshipit-source-id: 6d437ba2d404f1453e3bee61073e15cfcf43f8ae