Identify parameters for Pareto points in util
Summary:
The `pareto_frontier` utility currently returns a list of ObservationData corresponding to the pareto front. I'd like to be able to identify which points these are (i.e. what the corresponding parameters are). To do this I replaced the ObservationData with a full Observation that is thus able to track the features for each pareto point.

This makes `observation_features` a non-optional input. Previously its presence or not was used to decide whether to use model predictions or raw data to compute the Pareto front. I changed that to be an explicit flag.

I also added arm_names as an optional input, so we can directly see the arm name of the pareto points when using the utility on in-sample data.

Reviewed By: 2timesjay

Differential Revision: D26932046

fbshipit-source-id: 1944e1c71c289accb1d7c0cc95b256345c0efb07