Have pareto evaluator return index of Pareto points
Summary:
The frontier evaluator takes in a bunch of points and returns the subset that are Pareto non-dominated. It will be very helpful in a subsequent diff to know which of the input points were the Pareto ones (particularly so we can look up the arm name of the Pareto points).

This adds the index as a return from the frontier evaluator, and fixes all of the calls to use the new return signature.

Reviewed By: sdaulton

Differential Revision: D26931421

fbshipit-source-id: 5a8bf7a9c4157aa504e0c14c3eb082f9dbe7c956