Add plot for Pareto front of observed arms only
Summary:
The current Pareto plot uses the model to compute a Pareto front by maximizing model posterior mean.

It's also useful to have a Pareto plot that just uses the observed in-sample arms without requiring modeling and optimization. This can be useful (a) because it is alot faster, and (b) it doesn't require good model fit.

This diff implements that function. In doing so it makes a few tweaks to the current Pareto plot:

- Added arm name as an optional field in the Pareto plot input so that we can see the arm name in the hoverover for in-sample arms. It will continue getting a generic "Parameterization XX" for optimized arms.
- Shows all objective metrics on the overover, not just the two being plotted.

Reviewed By: 2timesjay, sdsingh

Differential Revision: D26932373

fbshipit-source-id: 6ff3951b3864c7870905850ce072944236f8039d