Enable mocking of functions in `ax/storage/sqa_store/save.py` and `load.py` (there was aliasing that got in the way) (#529)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/529

The way our init files are currently specified in `ax/storage/sqa_store` creates some shadowing: we have `ax.storage.sqa_store.save` the module (from `save.py` file) and the function ––`save_experiment`, aliased as `save` (or maybe even two of them, as the JSON save is aliased in `sqa_store/__init__.py` and then SQL `save_experiment` is aliased in `storage/__init__.py` –– not sure if aliasing gets chained or not, but this is beside the point). This makes it so that it's impossible do to something like `patch("ax.storage.sqa_store.save.<my_function>"`), because `ax.storage.sqa_store.save` is translated to a function (`save_experiment`) instead of a module and results in an error like:
```
AttributeError: <function save_experiment at 0x7f8a787eddc0> does not have the attribute 'my_function'
```

This was done for a good reason in D14997052 (https://github.com/facebook/Ax/commit/0bdbcbc53252e423898e9f088b740d52358eafc1), so putting up an RfC to collectively brainstorm how to fix. I can just not mock things in that file (although it is very helpful for testing and will also happen the same way for `load.py`), but I think we should at least all be aware of this behavior if not fix it.

Reviewed By: ldworkin

Differential Revision: D27128028

fbshipit-source-id: c10f4ab2d73ef027d81922f12f112d6f8aae9e69