ChoiceEncode transform
Summary:
Adds a `ChoiceEncode` transform that does essentially the same thing as `OrderedChoiceEncode`, but encodes all (not just ordered) choice parameters.

This can be used in place of one-hot encoding with the categorical kernel that dme65 has worked with in BoTorch.

This also moves the two transforms into a common `transforms.choice_encode` module. Technically this is backward compatibility breaking, but I don't know how many folks import the individual transforms manually, and if they do they can probably easily fix it.

The main thing to note for `ChoiceEncode` is that there is no rounding in the inverse transform, so `model.gen` will have to produce candidates that actually have integer values. This will be achieved by a custom optimization on the BoTorch end.

One could technically just not transform this to integers and just pass in the actual choices (e.g. strings), but that doesn't play well with the BoTorch/GPyTorch setup.

Reviewed By: bletham

Differential Revision: D27143251

fbshipit-source-id: b1f4d1e81c3b87901ea840bb95b76e19d64ac1a7