Back out "Load some more db_ids from database"
Summary: This diff broke a bunch of storage tests. It's not doing anything bad, it's just messing up our equality logic. I need to ship it in conjunction with some other changes I'm making, which will take care of that. But in the meantime, just reverting so I can get tests passing again.

Reviewed By: lena-kashtelyan

Differential Revision: D27170578

fbshipit-source-id: 368b34b71efa8f1dc9862e06bbd8580573c79db6