Speed up incomplete observation handling in observation_from_data
Summary:
With large datasets, observation_from_data can be very slow. Profiling revealed that most of the time was being spent in the pandas groupby filter to segment complete from incomplete observations. This applies a lambda to each row sequentially and is expensive for many rows.

The logic for separating out complete vs. incomplete observations was incorrect. It identified incomplete observations as those in which any column had a NaN. But SEM can be NaN when an observation is complete. The correct logic is to look for NaNs just in the feature_cols, which change this diff makes. Without that change, anytime you have NaN SEMs all of the data would end up being put into the `incomplete_df`. This had no impact on the actual outcome because the call to _observations_from_dataframe would end up being the same for both complete_df and incomplete_df, but now at least things go in the right dataframe.

With that logic fixed, I then added logic that checks if any rows are incomplete, just by checking if complete_feature_cols is different from feature_cols. If it isn't (and it won't be in most use cases, where we typically don't have incomplete observations), then we can entirely shortcut the expensive groupby and filter. This makes things much faster.

Reviewed By: sdsingh

Differential Revision: D27141512

fbshipit-source-id: 7b05bbc0e1f98486a7d49ca1fdf94d94961234d9