Improve performance of get_observed_pareto_frontiers plot utility
Summary:
This utility constructs a minimal torch model in order to transform data, search space, and optimization config (all of the things necessary to compute the Pareto front) into tensors. I want this utility to work for any search space, not just those with continuous parameters, so I chose a set of transforms that would be able to convert any search space to tensors: SearchSpaceToChoice converts the search space to a ChoiceParameter with a value for each arm, and then OneHot was applied to turn each value into a Float[0, 1].

The sklearn binarizer and encoder are very slow for a large number of arms. So this changes the minimal model to instead use OrderedChoiceEncode, which directly maps the arms to an Int[0, num_arms-1] parameter that is then turned into a float. This is much faster and makes the overhead for this transformation almost 0. In order to do this I just needed to tweak SearchSpaceToChoice to be able to produce an ordered ChoiceParameter rather than its usual unordered.

Reviewed By: sdsingh

Differential Revision: D27157324

fbshipit-source-id: 7d4d8988896b0bce7b6405b6f60b71821dc6af43