Unify handling of Pareto frontier and hypervolume calculations
Summary:
The utility for computing hypervolume calls `pareto_frontier` to compute the Pareto frontier, and then from that goes on to compute the hypervolume of those points. When doing so it repeats a bunch of transforming that was already done in pareto_frontier, so this diff changes pareto_frontier to directly return the transformed quantities so they don't have to be done again. There is no change to the logic in this, just a simplification of code.

The diff also makes a change to the logic for extract_objective_thresholds to require threshold to be specified either for all objectives or for None, and to always return an array of the length of outcomes. This is because the downstream consumer of this (pareto_frontier_evaluator) expects either None or a tensor of Y.shape[1], and will throw an error if len(objective_thresholds)!=len(outcomes), which the current logic permits. Because this function is currently only used by a Torch modelbridge, I also changed it to just directly return a tensor rather than an array.

Reviewed By: 2timesjay

Differential Revision: D27095987

fbshipit-source-id: 18d3614e1a77324b2176787774c9e7c13b9842b6