Slight refactor of GenerationStrategy generator runs
Summary:
Generator runs can be in one of three states:
* Unattached
* Attached to a trial
* Attached to a generation strategy

Currently, *in our database*, it's not possible for a generator run to be attached to *both* a trial and generation strategy. We [even have validation](https://fburl.com/diffusion/zgb8p5gq) against that. Instead, even if the "same" generator run should be on the trial / generation strategy, we save a copy of it on each.

Because we are now saving db_ids on user-facing classes (in order to pave the way for a storage overhaul), we need there to be a 1:1 mapping between user-facing instances and DB instances. Thus, we *also* need to copy generator runs and save different instances on trials and generator runs.

We're already doing this in a number of cases. In particular, whenever you call `trial.add_generator_run(gr)`, we make a copy of `gr` before attaching it to the trial.

So if your code looks like:
```
gr = gs.gen() # gr is attached to GS
trial.new_trial(gr) # a COPY of gr, call it gr2, is attached to trial
save_experiment(exp) # gr2 is saved to trial in DB
update_generation_strategy(
    generation_strategy=gs,
    new_generator_runs=[gr], # gr is saved to GS in DB
)
```
Then all works as expected. `gr` is saved to the DB and attached to `gs`, whereas `gr2` is attached to `trial`.

However, if your code looks like this, we're in trouble:

```
gr = gs.gen() # gr is attached to GS
trial.new_trial(gr) # a COPY of gr, call it gr2, is attached to trial
save_experiment(exp) # gr2 is saved to trial in DB
update_generation_strategy(
    generation_strategy=gs,
    new_generator_runs=trial.generator_runs, # gr2 is re-saved to GS in DB!
)
```

Our 1:1 mapping has been violated. The user-facing `gr2` instance maps to *two different* database objects, and the user-facing `gr` instance maps to none.

This isn't breaking anything now, but in order for my upcoming storage refactor to work, this 1:1 mapping needs to be intact.

To fix this, we need to add validation to make sure that  `update_generation_strategy` is only ever called with "fresh" generator runs that have not previously been saved, and we need to replace code paths that look like the second example with ones that look like the first example.

Reviewed By: lena-kashtelyan

Differential Revision: D27245364

fbshipit-source-id: 0cbd89320df029baf22ab56db52db7059f43902c