Update pre-save experiment validation
Summary:
Now that we're putting db_ids on our user-facing classes, we can update our pre-save experiment validation logic in a few ways:
- It currently takes as input the previous version of the experiment, if one exists. This isn't necessary -- we can instead provide just the db id of the previous version.
- We can raise clearer error messages that denote a few invalid states:
   * 1) changing the name of an experiment that has already been saved (bad),
   * 2) changing the name of an experiment that has already been saved to the name of an experiment that already exists in the database (even worse)
    * 3) creating a brand-new experiment and giving it the name of an experiment that already exists in the database

NOTE: This diff does change some behavior. Currently, if you load an experiment from the database, change its name, and then call `save_experiment`, we'll create a *brand-new clone* of the experiment with the new name and save it to the database. Now, if you try to do this, it will error. I think that's okay because the current behavior is a bit surprising. Thoughts?

Reviewed By: lena-kashtelyan

Differential Revision: D27280426

fbshipit-source-id: f5f67b225c74196f2e1b38fef3b4989cf3bfcda4