Introduce SearchSpaceDigest, pass more information to models (#533)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/533

Introduces `SearchSpaceDigest`, a container for lightweight representation of search space properties.

This is used for communicating between modelbridge and models. It contains additional information on `ordinal_features` and `categorical_features`, which is necessary information for models to be able to use discrete parameters directly (e.g. for categorical kernels). This will be utilized in subsequent diffs (currently the behavior for converting int parameters to tasks is unchanged, this will be modified in the future).

`SearchSpaceDigest` is an ephemeral object and not meant to be stored / serialized.

Reviewed By: bletham

Differential Revision: D27214304

fbshipit-source-id: ee972180e88af07576f040411205cc1c444724cb