Fix metric cloning
Summary:
Currently, if you don't define `clone()` on a metric subclass, it will fallback to the definition on the `Metric` base class, which returns an instance of the `Metric` base class.

For instance:
```
class MyMetric(Metric):
   pass

> m = MyMetric()
> type(m.clone())
Metric # but it should be MyMetric!
```

We defined `clone()` on all of our open source metrics (Branin, Factorial, etc) but we don't consistently do this on our FB-specific metrics (why would we?!).

Instead, let's change the behavior of the default `clone()` method so that it returns an instance of the subclass. This is just like what 2timesjay did with the AbstractData method `copy_structure_with_df` here: https://fburl.com/diffusion/d8zpr4wo

Reviewed By: stevemandala

Differential Revision: D27467306

fbshipit-source-id: 50eb7e0165121347cd60648ba758c57e94e5efc9