N TS arms should have total weight N
Summary:
As discussed recently with bletham, stevemandala, and Jakepodell, if Thompson Sampling returns N arms, the sum of the weights of all of those arms should be N, not 1 (as it currently is).

For instance, instead of [0.5, 0.25, 0.25], the weight should be [2, 1.5, 1.5]. Or, in the case of uniform weights, instead of [0.33, 0.33, 0.33], the weights should be [1, 1, 1].

In particular, this will help with the weird behavior we see in TS currently, where in an iteration batch with GPEI and TS arms, the GPEI arms always have higher weight (and therefore more segments) than the TS arms.

Reviewed By: bletham

Differential Revision: D27402744

fbshipit-source-id: 5e993b095bd094a8a3113a27130de6cf9f8c56e9