Hot fix for storage of ObjectiveThresholds
Summary:
In D27280549 (https://github.com/facebook/Ax/commit/4c53e921f52f96db91e4876e6866e02eb33b2ba1), I made a somewhat significant change to how our `save_experiment` logic works. Eventually this is going to make our storage layer a lot easier to reason about and more robust/scalable.

As part of that change, it's necessary that we always have a 1:1 relationship between user-facing objects and DB objects (and validation will complain if we don't).

This 1:1 relationship is necessary because we'll now be using a hidden `db_id` field on all user-facing objects to indicate which DB object they correspond to. We can then make use of SQLAlchemy's built in `merge` function to  synchronize the user-facing and DB representation and perform the necessary DB updates (rather than relying on our own custom logic to do this, as we currently are).

We should already be doing this everywhere, but there are a few places that slipped through the cracks (e.g. my recent fix in D27280534). Here is another one. When we do things like:

```
m1 = Metric()
m2 = Metric()

multi_obj = MultiObjective(m1, m2)

obj_thresh = ObjectiveThreshold(m1)
obj_thresh2 = ObjectiveThreshold(m2)
```

Then the user-facing `m1` and `m2` correspond to *two* different DB objects -- one for the multi objective and one for the objective threshold.

To fix this, we can just clone metrics that we pass into the ObjectiveThreshold constructor.

Reviewed By: stevemandala

Differential Revision: D27463284

fbshipit-source-id: 8ab4917c8a8218a828827896ab6d4381c5d51e91