Save TS weights in gen metadata
Summary:
In order to start tracking the “PTS success metric” that bletham and adamobeng came up with — namely, did Thompson Sampling find an arm with weight greater than that of the status quo? — we need to start saving to the DB the weights calculated by Thompson Sampling.

We already have the infrastructure to do this. Whenever you call `model.gen`, a dictionary of `gen_metadata` is returned and attached to the resulting `GeneratorRun` object.

Right now, Thompson Sampler's `gen_metadata` is empty. Instead, let's store a mapping of arms to the weights that TS calculated for them.

Reviewed By: stevemandala

Differential Revision: D27474771

fbshipit-source-id: 80926cf35880591cf2f66b176f5d79086165eafa