Fix save_experiment for AxQE
Summary:
I accidentally broke the behavior of `AxQE.save_experiment` recently -- see n531764 for an example.

The fix is to bypass some validation that I added that was checking equality of types. AxQE and ax.Experiment are not the same type, and that was throwing off my logic. However, for the purposes of our storage layer, they *are* actually identical. The extra stuff that is stored on the user-facing AxQE class does not get directly stored in the DB (with the exception of owners and tasks, but those are handled specially anyway).

Differential Revision: D27471724

fbshipit-source-id: e955f67cfbe5f8fb4b15df7d4c7f4db2f24c1fb4