Fix bug in modelbridge pareto utils
Summary:
In D26930981 (https://github.com/facebook/Ax/commit/1b3879bc3074168789a71be61bf37572944269cb) I fixed a bug where pareto visualizations and hypervolumes were incorrect if the metrics on the observations weren't in alphabetical order. Except there was a bug in my bugfix which meant that things still weren't correct if observations weren't in alphabetical order, unless there were only 2 in which case it was correct (which was unfortunately the case that I had tested). This fixes it for real this time.

dogscience

Reviewed By: sdsingh

Differential Revision: D27546376

fbshipit-source-id: 71b52fa19879ad74b0d4de080b89c22e58911b7d