Fix issue in copy_db_ids recursing on gen_metadata
Summary: As you see in D27573529 (https://github.com/facebook/Ax/commit/83968c6b3678f4f9bf6c5c5a258c00061df11b66), there's an issue in copy_db_ids right now when it gets into the bowels of `gen_metadata`. To recall, this is a recursive function that is meant to copy over db ids from one user facing object to another. The objects need to be identical for this to work, and they are! The issue is that in `gen_metadata` we have lists that look like [0, 'foobar']. While these lists are the same on both the source and target user facing object, Python is refusing to sort these lists for us. I'm not sure what the best solution is here, but for now, we can just stop recursing over this field (and related fields). There are no db_ids downstream of any of these fields, so its safe to skip and continue the recursion. I'm planning to revisit this function and figure out a better solution soon anyway (T88027346)

Reviewed By: lena-kashtelyan

Differential Revision: D27576747

fbshipit-source-id: 543173df20710f1df89fd56753888969baeef5df