]6.5n][RFC] Raise certain errors in unit tests only
Summary:
Curious to hear folks opinions on this. The issue is that I have some exceptions that I'd like to try/catch and gracefully continue on if they occur in production (since they're often not critical, and I don't want to totally break the user experience). But, I do want to be alerted of these failures when running unit tests.

I did some googling, and it seems like setting an environment variable in the unittest flow is a reasonable option here: https://stackoverflow.com/questions/25025928/how-can-a-piece-of-python-code-tell-if-its-running-under-unittest

But this feels a little hacky so I'm definitely open to alternative options if folks have other ideas!

Reviewed By: lena-kashtelyan

Differential Revision: D27592272

fbshipit-source-id: 6aa0bfbb9a861a2e999809ac14263f7fdc51d4b9