Respect immutable OC and SS when loading experiment
Summary: When `experiment.immutable_search_space_and_opt_config` is set, we don't save copies of SS and OC on each generator run. Therefore, we don't need to try to load these copies from the DB when we load the experiment. It wouldn't hurt anything, since the queries would just return empty lists, but they are expensive queries and they are wasteful. This will significantly speed up loading for experiments where this flag is set.

Reviewed By: lena-kashtelyan

Differential Revision: D27866442

fbshipit-source-id: 8bf4e6aaab7e6fa470a5f7107ae7a136dfe866c4