Add setter for immutable oc and ss
Summary: Adding a method for directly setting immutable_oc_and_ss on an experiment in the database. I don't imagine this will be used very much, but I've been using it during testing and figured I might as well check it in. It might come in handy if we want to migrate some old experiments in order to speed up their loading.

Reviewed By: lena-kashtelyan

Differential Revision: D27958006

fbshipit-source-id: c612ba224e9ac90f3e59be7c9ff2ce65e93352e3