Experiments created with AxClient should have immutable OC and SS
Summary:
As discussed [here](https://fb.workplace.com/groups/aeeng/permalink/937760730315957), currently, it's not supported behavior to allow a user to change the optimization config or search space in the middle of an experiment created by Scheduler or AxClient.

By actually enforcing this and not *allowing* the user to make any changes ([this error](https://fburl.com/diffusion/zabim6as) will be thrown if the user tries to manually reset the OC or SS on the experiment), we can stop saving copies of generator runs on these experiments, and significantly improve storage performance.

This is already enforced on the Scheduler and affects all Waveguide experiments. By enforcing on AxClient, it will start affecting ParamSweep and TRIP experiments.

As Balandat pointed out on my post, we might need to revisit this behavior when we productionize MOO in the Sweep and allow the user to adjust their objective thresholds in the middle. But as lena-kashtelyan replied [here](https://fb.workplace.com/groups/aeeng/permalink/937760730315957/?comment_id=937776883647675), it will be easy to adjust the behavior then to support that use case.

Reviewed By: lena-kashtelyan

Differential Revision: D27917947

fbshipit-source-id: 4a27977492e6389fed47e2324cf069e1f88a6fca